Initial commit!: initial code, files & folders
Unlike the first commit before which only added the files, this commit
adds more necessary files and changes that will be regarded as the first
commit.

changes:
Added description on Readme.md.
Minor changes to cmakelists.txt.
Added headers folder with header file to link main to header in folder
to see if it works without any other changes to cmake file.
Linked main to hello to test if it works

Breaking change:
Added hello.h in headers folder and inluded it in main but, try to
access hello from main by ctrl + click in main file results in
ERROR: File "headers/hello.h" not found in current source file's
directory or in build system paths.